Refactor data ingestion sources and add orchestrator
